SpreadsheetHateosHandlersSpreadsheetRouter extracted from SpreadsheetHateosHandlers
